add movie rating feature with external API integration and enhance movie management

+++ src/main/java/com/example/finalprogressreview/models/Movie.java
@@ -54,7 +54,7 @@ public class Movie {
         return releaseYear;
     }
 
-    public double getRating() {
+    public int getRating() {
         return rating;
     }
